Update linux_name_checker.py
I've updated and changed up the docstrings for each function and tidied up some of the string literals. In my humble opinion, doc strings are unsung heroes of python. They give the author an opportunity to describe their intent directly to the consumer of their code and it stays with the code so it can't be misplaced. There are all sorts of "standards" for what goes in a doc string, but the following is pretty universal:

```
def function(foo, bar, bar):
    """A one line sentence that describes what function does.

    :param int foo:  a comment describing what foo is
    :param str bar: a comment describing what bar is
    :param list baz: a comment describing what baz is
    :return: the type of object returned by function
    raises:
    - a list of exceptions and why each is raised by function
    Any other supporting documentation that the author feels
    will aid someone attempting to integrate this code into another
    project, or understand it so they can modify it competently.
   """
```